Refactor bctl server + get everything working with bzero-dev
