+++ bctl-go-daemon/Server/start.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+cd /bctl-server/Server/ && go run /bctl-server/Server/main.go -serviceURL=sid.bastionzero.com
